fix init problems
