+++ pkg/microservice/aslan/config/config.go
@@ -303,14 +303,6 @@ func DindImage() string {
 	return viper.GetString(setting.DindImage)
 }
 
-func MysqlDexDB() string {
-	return viper.GetString(setting.ENVMysqlDexDB)
-}
-
 func Features() string {
 	return viper.GetString(setting.FeatureFlag)
-}
-
-func MysqlUserDB() string {
-	return viper.GetString(setting.ENVMysqlUserDB)
 }
