+++ pkg/microservice/systemconfig/core/features/handler/features.go
@@ -48,7 +48,6 @@ func UpdateOrCreateFeature(c *gin.Context) {
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
 
 	if err != nil {
-
 		ctx.Err = fmt.Errorf("authorization Info Generation failed: err %s", err)
 		ctx.UnAuthorized = true
 		return
